+++ lib/pixel_app/common_widgets.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:pixel_preview/pixel_theme.dart';
-import 'package:pixel_preview/preview_widget.dart';
+import 'package:pixel_preview/pixel_preview/preview_widget.dart';
+import 'package:pixel_preview/utils/pixel_theme.dart';
 
 class EmptyState extends StatelessWidget {
   const EmptyState({
