+++ lib/pixel_app/components_builder.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:pixel_preview/preview_widget.dart';
+import 'package:pixel_preview/pixel_preview/preview_widget.dart';
 import 'package:pixel_preview/pixel_app/common_widgets.dart';
 
 class ComponentsBuilder extends StatefulWidget {
