+++ lib/pixel_app/screens_builder.dart
@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:pixel_preview/pixel_app/common_widgets.dart';
-import 'package:pixel_preview/pixel_theme.dart';
-import 'package:pixel_preview/preview_widget.dart';
+import 'package:pixel_preview/pixel_preview/preview_widget.dart';
+import 'package:pixel_preview/utils/pixel_theme.dart';
+import 'package:pixel_preview/utils/presets.dart';
+import 'package:pixel_preview/utils/sizes.dart';
 
 class ScreensBuilder extends StatefulWidget {
   final List<PixelPreview> screens;
