+++ lib/pixel_preview.dart
@@ -1,2 +1,3 @@
-export 'preview_widget.dart';
+export 'pixel_preview/preview_widget.dart';
 export 'pixel_app/pixel_app.dart';
+export 'utils/presets.dart';
