+++ lib/pixel_preview/frame_widget.dart
@@ -1,7 +1,8 @@
 import 'dart:math';
 
 import 'package:flutter/material.dart';
-import 'package:pixel_preview/pixel_theme.dart';
+import 'package:pixel_preview/utils/pixel_theme.dart';
+
 
 class FrameWidget extends StatefulWidget {
   final Widget child;
