+++ lib/pixel_preview/screenshot_thumbnail.dart
